style: consulta
