+++ frontend/paciente.html
@@ -7,7 +7,7 @@
     <!-- Folha de estilo: -->
     <link rel="stylesheet" href="./src/css/paciente.css" />
     <!-- Javascript: -->
-    <script src="./src/js/troca-conteudo-pacienteHTML.js" defer></script>
+    <script type="module" src="./src/js/paciente.js" defer></script>
   </head>
   <body>
     <header>
@@ -37,6 +37,8 @@
     </header>
     <main>
       <div class="mainDiv">
+
+            <!-- Menu Lateral -->
             <aside class="aside-dados-paciente">
             <h2 class="aside-title">DADOS</h2>
             <div class="aside-conteudo">
@@ -54,13 +56,18 @@
                 </ul>
             </nav>
             <div class="aside-buttons">
-                <button id="button-consulta">CONSULTA</button>
+                <!-- Bug: barrinha azul aparecendo embaixo do item, corrigir também os px gerais dos botões (Origamid aula) -->
+                <button id="button-consulta" class="aside-select">CONSULTA</button>
                 <button id="button-busca">BUSCA</button>
             </div>
             </div>
             </aside>
+            <!-- Fim Menu Lateral -->
+
+            <!-- Div/Container que estarão as abas de cada sessão: -->
             <div class="div-show-dados-paciente">
-                <div class="dados-pessoais-paciente container">
+                <!-- Sessão: Dados Pessoais -->
+                <div class="dados-pessoais-paciente container" style="display: none;">
                     <div class="area-caracteristicas-paciente">
                     <img
                         id="pacient-profile-img"
@@ -200,7 +207,9 @@
                     </p>
                     </div>
                 </div>
+                <!-- FIm Sessão: Dados Pessoais -->
 
+                <!-- Sessão: Histórico -->
                 <div class="container historico-paciente" style="display: none">
                     <div class="container-historico">
                         <div class="item-historico consulta">
@@ -409,6 +418,83 @@
                         </div>
                     </div>           
                 </div>
+                <!-- Fim Sessão: Histórico -->
+
+                <!-- Sessão: Consulta -->
+                <div class="consulta-paciente container" style="display: block;">
+                    <div data-consulta="divs" class="anamnese-div-container div-selecionada">
+
+                        <!-- Essa nova div utilizada para separar o conteúdo do botão, deverá ser repensada no refatoramento -->
+                        <div class="anamnese-div-container">
+                            <div class="anamnese-div-title">
+                                <h1>Anamnese</h1>
+                                <div class="title-div-informations">
+                                    <div class="area-label">
+                                        <label for="">Protocolo:</label>
+                                        <p>Dipirona</p>
+                                    </div>
+                                    <div class="area-label">
+                                        <label for="">Protocolo:</label>
+                                        <p>Dipirona</p>
+                                    </div>
+                                    <div class="area-label">
+                                        <label for="">Protocolo:</label>
+                                        <p>Dipirona</p>
+                                    </div>
+                                    <div class="area-label">
+                                        <label for="">Protocolo:</label>
+                                        <p>Dipirona</p>
+                                    </div>
+                                    <div class="area-label">
+                                        <label for="">Protocolo:</label>
+                                        <p>Dipirona</p>
+                                    </div>
+                                </div>
+                            
+                            </div>
+
+                            <div class="anemnese-space-bar"></div>
+                            
+                            <div class="ananmese-div-queixa">
+                                <h1>Queixa do Paciente</h1>
+                                <form action="" method="post">
+
+                                    <div>
+                                        <label for="queixa-atual">Atual:</label>
+                                        <input type="text" name="queixa-atual" id="">
+                                    </div>
+
+                                    <div>
+                                        <label for="queixa-progresso">Progresso:</label>
+                                        <input type="text" name="queixa-progresso" id="">
+                                    </div>
+
+                                </form>
+                            </div>
+                        </div>
+
+                        <button data-consulta="button" type="submit">Próximo</button>
+                    </div>
+
+                    <div data-consulta="divs" class="antecedentes-div-container">
+                        <h1>Antecedentes</h1>
+                        <button data-consulta="button" type="submit">Próximo</button>
+                    </div>
+
+                    <div data-consulta="divs" class="contra-indicacoes-div-container">
+                        <h1>Contra</h1>
+                        <button data-consulta="button" type="submit">Próximo</button>
+                    </div>
+
+                    <div data-consulta="divs" class="diagnostico-div-container">
+                        <h1>Diagnóstico</h1>
+                        <button data-consulta="button" type="submit">Próximo</button>
+                    </div>
+
+                </div>
+                <!-- FIm Sessão: Consulta -->
+
+
             </div>
        </div>
     </main>
