+++ frontend/src/css/paciente.css
@@ -329,4 +329,117 @@ main {
     display: none;
 }
 /* fim area de historico */
+
+
+/* Consulta: */
+.consulta-paciente {
+    height: 100%;
+}
+
+[data-consulta="divs"] {
+    display: none;
+} 
+
+.div-selecionada {
+    display: block;
+    height: 100%;
+}
+
+.anamnese-div-container {
+    /* background: red; */
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    padding: 20px;
+    gap: 35px;
+    width: 100%;
+}
+
+.anamnese-div-title {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    width: 100%;
+    
+}
+.anamnese-div-title h1 {
+    background: var(--color-blue);
+    padding: 0px 60px;
+    border-radius: 10px;
+    color: var(--color-white);
+    align-self: center;
+}
+.title-div-informations {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    justify-items: center;
+    gap: 10px;
+}
+
+.anemnese-space-bar {
+    height: 4px;
+    background: var(--color-mediumAqua);
+    width: 100%;
+    border-radius: 10px;
+}
+
+.ananmese-div-queixa {
+    display: flex;
+    flex-direction: column;
+    /* background: red; */
+    width: 100%;
+    align-items: center;
+    gap: 20px;
+}
+.ananmese-div-queixa h1 {
+    font-size: 1.8rem;
+}
+.ananmese-div-queixa form {
+    display: grid;
+    width: 100%;
+    gap: 8px;
+}
+.ananmese-div-queixa form div {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 5px;
+}
+
+.ananmese-div-queixa form div input {
+    height: 30px;
+    border-radius: 12px;
+    border: none;
+    background: var(--color-light-gray);
+    width: 100%;
+}
+.anamnese-div-container button {
+    align-self: flex-end;
+    position: relative;
+    bottom: 0;
+
+    padding: 5px 60px;
+    border: none;
+    background: var(--color-mediumAqua);
+    border-radius: 10px;
+    color: var(--color-white);
+    font-weight: 600;
+    font-size: 1.2rem;
+    cursor: pointer;
+    transition: .3s;
+
+}
+.anamnese-div-container button:hover {
+    background: var(--color-blue);
+}
+/* Fim Consulta */
+
+
 /* fim area de mostra de dados */
+
+/* Por enquanto: */
+.area-label {
+    display: flex;
+    gap: 5px;
+}
