+++ frontend/src/js/paciente.js
@@ -1,0 +1,5 @@
+import initTrocaConteudo from './modules/troca-conteudo-pacienteHTML.js';
+import initTrocaDivConsulta from './modules/troca-div-consulta.js';
+
+initTrocaConteudo();
+initTrocaDivConsulta();
